fix(config): extraneous keys and values overwrite explicit matching key values
