+++ api/src/test/java/io/artframework/util/ConfigUtilTest.java
@@ -36,7 +36,7 @@ import java.io.File;
 import static org.assertj.core.api.Assertions.*;
 
 @SuppressWarnings("ALL")
-class ConfigUtilTest {
+public class ConfigUtilTest {
 
     @Nested
     @DisplayName("toConfigString()")
@@ -66,143 +66,6 @@ class ConfigUtilTest {
 
             assertThat(ConfigUtil.toConfigString(ConfigUtil.getConfigFields(ArrayConfig.class)))
                     .isEqualTo("array=String..., foo=0");
-        }
-    }
-
-    @Nested
-    class ConfigMap {
-
-        @Test
-        @DisplayName("should load all fields including superclass")
-        public void shouldLoadAllFields() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .hasSizeGreaterThanOrEqualTo(4)
-                    .containsKeys(
-                            "parent_field",
-                            "required",
-                            "default_field",
-                            "all_annotations"
-                    )
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should load required annotation")
-        public void shouldLoadRequiredAttribute() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .extractingByKey("required")
-                    .extracting(ConfigFieldInformation::required)
-                    .isEqualTo(true)
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should load description annotation")
-        public void shouldLoadDescriptionAttribute() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .extractingByKey("default_field")
-                    .extracting(ConfigFieldInformation::description)
-                    .isEqualTo(new String[]{"World to teleport the player to."})
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should load default value")
-        public void shouldLoadDefaultValue() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .extractingByKey("default_field")
-                    .extracting(ConfigFieldInformation::defaultValue)
-                    .isEqualTo("world")
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should load required field with default value")
-        public void shouldLoadRequiredDefaultValue() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .extractingByKey("all_annotations")
-                    .extracting(ConfigFieldInformation::defaultValue, ConfigFieldInformation::description)
-                    .contains(2.0d, new String[]{"Required field with default value."})
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should load nested config objects")
-        public void shouldLoadNestedObjects() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .extractingByKey("nested.nested_field")
-                    .extracting(ConfigFieldInformation::description, ConfigFieldInformation::defaultValue)
-                    .contains(new String[]{"nested config field"}, "foobar")
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should not load nested object fields")
-        public void shouldNotAddNestedBase() {
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .doesNotContainKey("nested")
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should ignore fields without an annotation")
-        public void shouldIgnoredIgnored() {
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .doesNotContainKeys("ignored", "no_annotations")
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should load field position annotation")
-        public void shouldLoadFieldPosition() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(TestConfig.class))
-                    .extractingByKeys("required", "parent_field")
-                    .extracting(ConfigFieldInformation::position)
-                    .contains(1, 0)
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should throw if same field position is found")
-        public void shouldThrowExceptionForSamePosition() {
-
-            assertThatExceptionOfType(ConfigurationException.class)
-                    .isThrownBy(() -> ConfigUtil.getConfigFields(SamePositionConfig.class))
-                    .withMessageContaining("same position");
-        }
-
-        @Test
-        @DisplayName("should throw if declared config field is final")
-        void shouldThrowIfConfigOptionIsFinal() {
-
-            assertThatExceptionOfType(ConfigurationException.class)
-                    .isThrownBy(() -> ConfigUtil.getConfigFields(FinalConfig.class))
-                    .withMessageContaining("final field");
-        }
-
-        @Test
-        @DisplayName("should load all fields if the class is annotated")
-        void shouldLoadAllFieldsInAnnotatedClass() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(AnnotatedClass.class))
-                    .containsOnlyKeys("foo", "bar")
-            ).doesNotThrowAnyException();
-        }
-
-        @Test
-        @DisplayName("should load array field")
-        void shouldLoadArrayField() {
-
-            assertThatCode(() -> assertThat(ConfigUtil.getConfigFields(ArrayConfig.class))
-                    .containsOnlyKeys("foo", "array")
-            ).doesNotThrowAnyException();
         }
     }
